dj_api_luohu/requirements.txt
altgraph==0.17.2
APScheduler==3.9.1
asgiref==3.5.2
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
djangorestframework==3.13.1
drf-yasg==1.21.3
et-xmlfile==1.1.0
future==0.18.2
ghp-import==2.0.2
idna==3.2
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.3.0
mkdocs-material==8.2.9
mkdocs-material-extensions==1.0.3
numpy==1.21.2
openpyxl==3.0.9
packaging==21.3
pandas==1.3.3
pefile==2021.9.3
pinyin==0.4.0
Pygments==2.11.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pymdown-extensions==9.3
pymssql==2.2.5
pyodbc==4.0.32
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
pyyaml-env-tag==0.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.6
watchdog==2.1.7
win32-setctime==1.0.3
zipp==3.8.0
